Fix incorrect use of prop-types.any for intersection
The code generated to import a type from another file will
default to prop-types.any if the call to require() fails.
This will break if the imported type is used in an intersection,
because the generated code for that case will merge the
prop-types at runtime via Object.assign. For this case, we
check for prop-types.any and replace with '{}'.