Export objects, not shapes - shape-ify on import side
To properly handle intersections for imported types at runtime, we
need to export these types as objects, not shapes.

This is an intermediate commit, fixing the exports and only some
pieces of the import side while updating the tests.

Interesting, we still need to default to proptypes.any on failed
imports instead of {}. We don't know how to imported type would be
used, and proptypes.any will work in all contexts. For this reason,
we still need to check (as we do currently) for proptypes.any in
the runtime merge code, i.e. in the arguments to Object.assign.

Missing:

* Do not shape-ify variables in the runtime merge code (transpile time)
* Do not shape-ify functions (check at runtime)