Add support for instanceOf on possible classes
If we see a variable and it's not an imported type, a local type
alias or in a fixed set of items (Function, Object,..) we
may assume that the variable refers to a class constructor.
We add a runtime check if the variable is indeed a function,
in which case we add an instanceOf proptype. Else, we use the 'any'
proptype.

Also see #6, from which this idea was taken:

https://github.com/brigand/babel-plugin-flow-react-proptypes/pull/6/